+++ Source/Modules/Tests/Gapotchenko.FX.Numerics.Tests/BitOperationTests.cs
@@ -6,13 +6,13 @@ namespace Gapotchenko.FX.Numerics.Tests;
 public class BitOperationTests
 {
     [TestMethod]
-    // Convention.
+    // Contract
     [DataRow(0x0U, 0)]
-    // Sanity checks.
+    // Sanity checks
     [DataRow(0x1U, 0)]
     [DataRow(0x2U, 1)]
     [DataRow(uint.MaxValue, 31)]
-    // Verification.
+    // Verification
     [DataRow(0x20U, 5)]
     [DataRow(0x7fU, 6)]
     [DataRow(0x80U, 7)]
@@ -24,13 +24,13 @@ public class BitOperationTests
     }
 
     [TestMethod]
-    // Convention.
+    // Contract
     [DataRow(0x0UL, 0)]
-    // Sanity checks.
+    // Sanity checks
     [DataRow(0x1UL, 0)]
     [DataRow(0x2UL, 1)]
     [DataRow(ulong.MaxValue, 63)]
-    // Verification.
+    // Verification
     [DataRow(0x20UL, 5)]
     [DataRow(0x7fUL, 6)]
     [DataRow(0x80UL, 7)]
@@ -43,6 +43,46 @@ public class BitOperationTests
     public void BitOperations_Log2_UInt64(ulong value, int log)
     {
         Assert.AreEqual(log, BitOperations.Log2(value));
+    }
+
+    [TestMethod]
+    // Sanity checks
+    [DataRow(0x0U, 32)]
+    [DataRow(0x1U, 31)]
+    [DataRow(0x2U, 30)]
+    [DataRow(uint.MaxValue, 0)]
+    // Verification
+    [DataRow(0x20U, 26)]
+    [DataRow(0x7fU, 25)]
+    [DataRow(0x80U, 24)]
+    [DataRow(0xffU, 24)]
+    [DataRow(0x100U, 23)]
+    [DataRow(0x7fffffffU, 1)]
+    [DataRow(0x80000000U, 0)]
+    public void BitOperations_LeadingZeroCount_UInt32(uint value, int count)
+    {
+        Assert.AreEqual(count, BitOperations.LeadingZeroCount(value));
+    }
+
+    [TestMethod]
+    // Sanity checks
+    [DataRow(0x0UL, 64)]
+    [DataRow(0x1UL, 63)]
+    [DataRow(0x2UL, 62)]
+    [DataRow(ulong.MaxValue, 0)]
+    // Verification
+    [DataRow(0x20UL, 58)]
+    [DataRow(0x7fUL, 57)]
+    [DataRow(0x80UL, 56)]
+    [DataRow(0xffUL, 56)]
+    [DataRow(0x100UL, 55)]
+    [DataRow(0xffffffffUL, 32)]
+    [DataRow(0x100000000UL, 31)]
+    [DataRow(0x7fffffffffffffffUL, 1)]
+    [DataRow(0x8000000000000000UL, 0)]
+    public void BitOperations_LeadingZeroCount_UInt64(ulong value, int count)
+    {
+        Assert.AreEqual(count, BitOperations.LeadingZeroCount(value));
     }
 
     [TestMethod]